feat: svg image support
