+++ app/build.gradle.kts
@@ -142,6 +142,7 @@ dependencies {
     implementation(Libraries.coil)
     implementation(Libraries.coilGif)
     implementation(Libraries.coilCompose)
+    implementation(Libraries.coilSvg)
 
     /** lifecycle **/
     // ViewModel
